Crate aws_smithy_http_client

HTTP client implementation for smithy-rs generated code.

Crate Features

default-client: Enable default HTTP client implementation (based on hyper 1.x).

rustls-ring: Enable TLS provider based on rustls using ring as the crypto provider

rustls-aws-lc: Enable TLS provider based on rustls using aws-lc as the crypto provider

rustls-aws-lc-fips: Same as rustls-aws-lc feature but using a FIPS compliant version of aws-lc

s2n-tls: Enable TLS provider based on s2n-tls using aws-lc as the crypto provider.

hyper-014: (Deprecated) HTTP client implementation based on hyper-0.14.x.

test-util: Enables utilities for unit tests. DO NOT ENABLE IN PRODUCTION.

Modules

hyper_014Deprecated

Legacy HTTP and TLS connectors that use hyper 0.14.x and rustls.

proxy

Proxy configuration Proxy configuration for HTTP clients

tls

TLS connector(s)

Structs

Builder

Builder for a hyper-backed HttpClient implementation.

Connector

HttpConnector used to make HTTP requests.

ConnectorBuilder

Builder for Connector.

HttpClientError

HTTP client errors

Functions

default_connector

Given HttpConnectorSettings and an SharedAsyncSleep, create a SharedHttpConnector from defaults depending on what cargo features are activated.