aws_smithy_http_client

Module hyper_014

👎Deprecated: hyper 0.14.x support is deprecated, please migrate to 1.x client

Legacy HTTP and TLS connectors that use hyper 0.14.x and rustls.

Structs

HyperClientBuilder

Builder for a hyper-backed HttpClient implementation.

HyperConnector

HttpConnector that uses hyper_0_14 to make HTTP requests.

HyperConnectorBuilder

Builder for HyperConnector.

Functions

default_client

Creates a hyper-backed HTTPS client from defaults depending on what cargo features are activated.

default_connector

Given HttpConnectorSettings and an SharedAsyncSleep, create a SharedHttpConnector from defaults depending on what cargo features are activated.