aws_smithy_http_client

Struct ConnectorBuilder

pub struct ConnectorBuilder<Tls = TlsUnset> { /* private fields */ }

Builder for Connector.

Implementations

impl ConnectorBuilder<TlsUnset>

pub fn tls_provider( self, provider: Provider, ) -> ConnectorBuilder<TlsProviderSelected>

Set the TLS implementation to use for this connector

impl<Any> ConnectorBuilder<Any>

pub fn sleep_impl(self, sleep_impl: impl AsyncSleep + 'static) -> Self

Set the async sleep implementation used for timeouts

Calling this is only necessary for testing or to use something other than default_async_sleep.

pub fn set_sleep_impl( &mut self, sleep_impl: Option<SharedAsyncSleep>, ) -> &mut Self

Set the async sleep implementation used for timeouts

Calling this is only necessary for testing or to use something other than default_async_sleep.

pub fn connector_settings( self, connector_settings: HttpConnectorSettings, ) -> Self

Configure the HTTP settings for the HyperAdapter

pub fn set_connector_settings( &mut self, connector_settings: Option<HttpConnectorSettings>, ) -> &mut Self

Configure the HTTP settings for the HyperAdapter

pub fn enable_tcp_nodelay(self, nodelay: bool) -> Self

Configure SO_NODELAY for all sockets to the supplied value nodelay

pub fn set_enable_tcp_nodelay(&mut self, nodelay: bool) -> &mut Self

Configure SO_NODELAY for all sockets to the supplied value nodelay

pub fn set_interface<S: Into<String>>(&mut self, interface: S) -> &mut Self

Sets the value for the SO_BINDTODEVICE option on this socket.

If a socket is bound to an interface, only packets received from that particular interface are processed by the socket. Note that this only works for some socket types (e.g. AF_INET sockets).

On Linux it can be used to specify a VRF, but the binary needs to either have CAP_NET_RAW capability set or be run as root.

This function is only available on Android, Fuchsia, and Linux.

pub fn proxy_config(self, config: ProxyConfig) -> Self

Configure proxy settings for this connector

This method allows you to set explicit proxy configuration for the HTTP client. The proxy configuration will be used to determine whether requests should be routed through a proxy server or connect directly.

Examples

use aws_smithy_http_client::{Connector, proxy::ProxyConfig, tls};

let proxy_config = ProxyConfig::http("http://proxy.example.com:8080")?;
let connector = Connector::builder()
    .proxy_config(proxy_config)
    .tls_provider(tls::Provider::Rustls(tls::rustls_provider::CryptoMode::AwsLc))
    .build();

pub fn set_proxy_config(&mut self, config: Option<ProxyConfig>) -> &mut Self

Configure proxy settings for this connector

This is the mutable version of proxy_config.

pub fn pool_idle_timeout<D>(self, val: D) -> Self
where D: Into<Option<Duration>>,

Set an optional timeout for idle sockets being kept-alive.

Pass None to disable timeout.

Defaults to Hyperโs default timeout, which is currently 90 seconds - see hyper_util::client::legacy::Builder::pool_idle_timeout, but unlike that function, there is no need to call pool_timer yourself.

Examples

use aws_smithy_http_client::{Connector, tls};
use std::time::Duration;

let connector = Connector::builder()
    .pool_idle_timeout(Duration::from_secs(30))
    .tls_provider(tls::Provider::Rustls(tls::rustls_provider::CryptoMode::AwsLc))
    .build();

pub fn set_pool_idle_timeout( &mut self, val: Option<Option<Duration>>, ) -> &mut Self

Set an optional timeout for idle sockets being kept-alive.

Pass None to use Hyperโs default timeout, Some(None) to disable timeouts.

This is the mutable version of pool_idle_timeout.

Examples

use aws_smithy_http_client::{Connector, tls};
use std::time::Duration;

let mut connector = Connector::builder();
connector
    .set_pool_idle_timeout(Some(Some(Duration::from_secs(30))));
connector
    .tls_provider(tls::Provider::Rustls(tls::rustls_provider::CryptoMode::AwsLc))
    .build();

impl ConnectorBuilder<TlsProviderSelected>

pub fn build(self) -> Connector

Build a Connector that will use the default DNS resolver implementation.

pub fn tls_context(self, ctx: TlsContext) -> Self

Configure the TLS context

pub fn set_tls_context(&mut self, ctx: TlsContext) -> &mut Self

Configure the TLS context

pub fn build_with_resolver<R: ResolveDns + Clone + 'static>( self, resolver: R, ) -> Connector

Build a Connector that will use the given DNS resolver implementation.